bot/requirements.txt
discord.py>=2.3,<3
anthropic>=0.40,<1
aiohttp>=3.9,<4
httpx>=0.27,<1
